refactor: put misc pkg init function to separate file
